KL-147/refactor: separate file for useProductQuery hook
useProductQuery 훅을 별도 파일로 분리하였습니다.
